minor fix & wallet Package 추가

+++ spring-backend/src/main/java/com/blooming/blockchain/springbackend/points/controller/PointsController.java
@@ -94,7 +94,7 @@ public class PointsController {
         return ResponseEntity.ok(PointsResponse.builder()
             .success(true)
             .mainPointsReceived(mainPointsReceived)
-            .conversionRate(PointTransaction.getSubToMainRatio())
+            .conversionRate((int) PointTransaction.getSubToMainRatio())
             .message("Points converted successfully")
             .build());
     }
